Implementation of basic methods to retrieve data from sncf API
Getting data from the remote API will be one of the key element of
this project. That's why two methods has been implemented.

The first one is very basic and just performs a get request by creating
the url and sending data. It should be the only method to actually
performs GET request.

The second one was created in order to fill the missing functionality of
the navitia API. In fact, we can't retrieve all records with one request.
That's why this method makes multiple requests in order to get all data
and concat them as one final response.