Implementing a Storage process and filter
As pointed out in a previous commit, retrieving all stations takes time
(more than 3.5 seconds), which is too much. Secondly it uses 4 requests
each time that we could easily spare.

That's why a storage process was implementing. When the resource is
requested for the first time, the data are retrieved from the remote
API. The second time, the path to this resource is given as a result.

Thanks to this, we dropping from 3.5sec to 0.2sec.

A filter function will allow to filter results to match parameters. For
now it is only filtering on wether GPS coordinates are present and
correct (different from 0.0/0.0)